+++ database/migrations/1772419332000_create_auth_access_tokens_table.ts
@@ -5,10 +5,10 @@ export default class extends BaseSchema {
 
   async up() {
     this.schema.createTable(this.tableName, (table) => {
-      table.bigIncrements('id').primary()
+      table.increments('id').primary()
 
       table
-        .bigInteger('tokenable_id')
+        .integer('tokenable_id')
         .unsigned()
         .notNullable()
         .references('id')
